Move MainActivity to presentation package
Suppress MagicNumber in HomeViewModule
